Revert "rename examples to test_examples.py"
This reverts commit dea60b70fd2f28162f074cdb7f39ed835d581ccf.
