+++ tests/e2e/default_openaispec_response_format.py
@@ -1,6 +1,6 @@
 import litserve as ls
 from litserve import OpenAISpec
-from litserve.test_examples.openai_spec_example import TestAPIWithStructuredOutput
+from litserve.examples.openai_spec_example import TestAPIWithStructuredOutput
 
 if __name__ == "__main__":
     server = ls.LitServer(TestAPIWithStructuredOutput(), spec=OpenAISpec())
