+++ tests/e2e/default_spec.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from litserve.test_examples.openai_spec_example import TestAPI
+from litserve.examples.openai_spec_example import TestAPI
 from litserve.specs.openai import OpenAISpec
 
 import litserve as ls
